Fixed _adc_res bug. Made some changes to the API.

+++ src/LM35IC.cpp
@@ -4,7 +4,7 @@
 namespace LM35{
 
     LM35IC::LM35IC(uint8_t pin, sensors sensor, units unit, double Vcc, uint8_t ADC_bits): 
-        _pin(pin), _sensor(sensor), _unit(unit), _adc_res(Vcc / pow(2.0, ADC_bits)){ 
+        _pin(pin), _sensor(sensor), _unit(unit), _adc_res(Vcc / pow(2.0, static_cast<double>(ADC_bits))){ 
 
         };
